refactor name and junit scope change
