+++ G26NewProject/AleksandrovProj/pom.xml
@@ -10,14 +10,13 @@
   <groupId>AlexsandrovProj</groupId>
   <artifactId>AlexsandrovProj</artifactId>
   <version>1.0-SNAPSHOT</version>
-  <name>AlexsandrovProj</name>
+  <name>AleksandrovProj</name>
   <url>http://maven.apache.org</url>
   <dependencies>
     <dependency>
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
       <version>3.8.1</version>
-      <scope>test</scope>
     </dependency>
   </dependencies>
 </project>
