+++ G26NewProject/AleksandrovProj/src/main/java/AleksandrovProj/App.java
@@ -1,4 +1,4 @@
-package AlexsandrovProj;
+package AleksandrovProj;
 
 /**
  * Hello world!
